+++ src/draggers/SoJackDragger.cpp
@@ -187,7 +187,7 @@ SoJackDragger::SoJackDragger(void)
   if (SO_KIT_IS_FIRST_INSTANCE()) {
     SoInteractionKit::readDefaultParts("jackDragger.iv",
                                        JACKDRAGGER_draggergeometry,
-                                       strlen(JACKDRAGGER_draggergeometry));
+                                       (int)strlen(JACKDRAGGER_draggergeometry));
   }
 
   SO_KIT_ADD_FIELD(rotation, (SbRotation(SbVec3f(0.0f, 0.0f, 1.0f), 0.0f)));
