+++ src/draggers/SoScaleUniformDragger.cpp
@@ -140,7 +140,7 @@ SoScaleUniformDragger::SoScaleUniformDragger(void)
   if (SO_KIT_IS_FIRST_INSTANCE()) {
     SoInteractionKit::readDefaultParts("scaleUniformDragger.iv",
                                        SCALEUNIFORMDRAGGER_draggergeometry,
-                                       strlen(SCALEUNIFORMDRAGGER_draggergeometry));
+                                       (int)strlen(SCALEUNIFORMDRAGGER_draggergeometry));
   }
   SO_KIT_ADD_FIELD(scaleFactor, (1.0f, 1.0f, 1.0f));
