+++ src/draggers/SoTransformBoxDragger.cpp
@@ -240,7 +240,7 @@ SoTransformBoxDragger::SoTransformBoxDragger(void)
   if (SO_KIT_IS_FIRST_INSTANCE()) {
     SoInteractionKit::readDefaultParts("transformBoxDragger.iv",
                                        TRANSFORMBOXDRAGGER_draggergeometry,
-                                       strlen(TRANSFORMBOXDRAGGER_draggergeometry));
+                                       (int)strlen(TRANSFORMBOXDRAGGER_draggergeometry));
   }
 
   SO_KIT_ADD_FIELD(rotation, (SbRotation(SbVec3f(0.0f, 0.0f, 1.0f), 0.0f)));
